+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/assetedition/screen/AssetEditManagement.kt
@@ -1,4 +1,4 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.screen
+package com.fintern.ourbudgeting.ui.assetmanagement.assetedition.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -11,10 +11,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.fintern.ourbudgeting.R
-import com.fintern.ourbudgeting.ui.assetmanagement.component.AssetEditBody
-import com.fintern.ourbudgeting.ui.assetmanagement.component.AssetEditTopAppBar
-import com.fintern.ourbudgeting.ui.assetmanagement.component.AssetTitle
-import com.fintern.ourbudgeting.ui.assetmanagement.component.AssetTopSection
+import com.fintern.ourbudgeting.ui.assetmanagement.assetedition.component.AssetEditBody
+import com.fintern.ourbudgeting.ui.assetmanagement.assetedition.component.AssetEditTopAppBar
+import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetTitle
+import com.fintern.ourbudgeting.ui.assetmanagement.common.component.AssetTopSection
 import com.fintern.ourbudgeting.ui.theme.OurBudgetingTheme
 
 @Composable
