+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/assetstatistics/component/StatisticAssetChartTitle.kt
@@ -1,4 +1,4 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.component
+package com.fintern.ourbudgeting.ui.assetmanagement.assetstatistics.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
