+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/assettypeaddition/component/AddAssetTopAppBar.kt
@@ -1,8 +1,9 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.component
+package com.fintern.ourbudgeting.ui.assetmanagement.assettypeaddition.component
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
@@ -33,7 +34,8 @@ fun AddAssetTopAppBar(
             IconButton(onClick = onActionClick) {
                 Icon(
                     painter = painterResource(R.drawable.ic_arrowback),
-                    contentDescription = stringResource(R.string.graph)
+                    contentDescription = stringResource(R.string.graph),
+                    tint = Color.Black
                 )
             }
         }
