+++ app/src/main/java/com/fintern/ourbudgeting/ui/assetmanagement/common/component/AssetBody.kt
@@ -1,4 +1,4 @@
-package com.fintern.ourbudgeting.ui.assetmanagement.component
+package com.fintern.ourbudgeting.ui.assetmanagement.common.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -9,13 +9,12 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.fintern.ourbudgeting.R
 
 @Composable
-fun AssetTitle(
+fun AssetBody(
     modifier: Modifier = Modifier,
     name: String,
     amount: Long
@@ -29,28 +28,26 @@ fun AssetTitle(
         Text(
             text = name,
             style = MaterialTheme.typography.titleMedium,
-            fontWeight = FontWeight.Bold
         )
         Text(
             text = stringResource(R.string.amount_won, amount),
             style = MaterialTheme.typography.titleMedium,
-            fontWeight = FontWeight.Bold
         )
     }
 }
 
 @Composable
 @Preview(showBackground = true)
-fun CashTitlePreview() {
+fun CashBodyPreview() {
     MaterialTheme {
-        AssetTitle(name = "현금", amount = 0)
+        AssetBody(name = "현금", amount = 0)
     }
 }
 
 @Composable
 @Preview(showBackground = true)
-fun BankTitlePreview() {
+fun BankBodyPreview() {
     MaterialTheme {
-        AssetTitle(name = "은행", amount = 0)
+        AssetBody(name = "은행", amount = 0)
     }
 }
